moved zero_like to numpy utils and fixed its bug with number types
